Rename labeling.py to iclabel.py
