+++ src/mnelab/utils/__init__.py
@@ -3,7 +3,7 @@
 # License: BSD (3-clause)
 
 from mnelab.utils.dependencies import have
-from mnelab.utils.labeling import run_iclabel
+from mnelab.utils.iclabel import run_iclabel
 from mnelab.utils.syntax import PythonHighlighter, format_code
 from mnelab.utils.utils import (
     Montage,
